refactor: added config directory to house configuration files
