+++ backend/helpers/commonHelpers.js
@@ -1,4 +1,4 @@
-const db_connection = require("../db_connection");
+const db_connection = require("../config/databaseConfig");
 
 const promiseBasedQuery = (query, values) => {
     /**
