+++ backend/helpers/groupControllerHelpers.js
@@ -35,6 +35,8 @@ const createGroupsOCEAN = () => {
 
 }
 
+
+
 const groupFormationStrategies = {
     /**
      *
